Add Leave One Out Cross validation functionality
Add a findBestWindow function to BaseDTW for finding the best window
using LOO CV over the training data
Change the code for initialization of the DynamicTimeWarping object in
NormalDTW, LuckyDTW and HeuristicDTW and call the function for finding
best window if the corresponding argument is provided.
Add a condition for accomodating the new argument value
Change code in the command line options help text and the default value
for window parameter.